requirements.txt
flask==0.10.1
flask_mongoengine==0.7.0
flask-security
py-bcrypt
pil
